+++ packages/engine-multi/test/errors.test.ts
@@ -1,5 +1,5 @@
 import test from 'ava';
-import createEngine from '../src/engine';
+import createEngine, { EngineOptions } from '../src/engine';
 import { createMockLogger } from '@openfn/logger';
 import { WORKFLOW_ERROR } from '../src/events';
 
@@ -8,7 +8,7 @@ let engine;
 test.before(async () => {
   const logger = createMockLogger('', { level: 'debug' });
 
-  const options = {
+  const options: EngineOptions = {
     logger,
     repoDir: '.',
     autoinstall: {
@@ -16,6 +16,7 @@ test.before(async () => {
       handleIsInstalled: async () => true,
     },
     maxWorkers: 1,
+    memoryLimitMb: 200,
   };
 
   // This uses the real runtime and real worker
@@ -64,6 +65,31 @@ test('syntax error: illegal throw', (t) => {
       // compilation happens in the main thread
       t.is(evt.threadId, '-');
       t.is(evt.message, 'z: Unexpected token (1:9)');
+      done();
+    });
+  });
+});
+
+test('syntax error: oom error', (t) => {
+  return new Promise((done) => {
+    const plan = {
+      id: 'a',
+      jobs: [
+        {
+          expression: `export default [(s) => {
+              s.a = [];
+              while(true) {
+                s.a.push(new Array(1e6).fill("oom"));
+              }
+              return s;
+            }]`,
+        },
+      ],
+    };
+
+    engine.execute(plan).on(WORKFLOW_ERROR, (evt) => {
+      t.is(evt.type, 'OOMError');
+      t.is(evt.severity, 'kill');
       done();
     });
   });