+++ packages/runtime/ava.config.cjs
@@ -1,0 +1,7 @@
+const baseConfig = require('../../ava.config');
+
+module.exports = {
+  ...baseConfig,
+
+  files: ['!test/memory.test.ts'],
+};